fix: delete test resources

+++ k8s/test.sh
@@ -89,7 +89,7 @@ __safe_migration_rollout() {
 
 upgrade_alembic_migrations() {
     echo "::group::Upgrade alembic migrations for test"
-    if [ $KUBERNETES_DEPLOYMENT_NAME != "refinery-gateway" ] && [ $KUBERNETES_DEPLOYMENT_NAME != "gates-gateway" ] && [ $KUBERNETES_DEPLOYMENT_NAME != "hosted-inference-api" ]; then
+    if [ $KUBERNETES_DEPLOYMENT_NAME != "refinery-gateway" ] && [ $KUBERNETES_DEPLOYMENT_NAME != "hosted-inference-api" ]; then
         kubectl apply --kustomize apps/${REFINERY_DEPLOYMENT_NAME}/test
         __safe_migration_rollout test-${REFINERY_DEPLOYMENT_NAME}
         echo "Applied test-${REFINERY_DEPLOYMENT_NAME} deployment"
@@ -170,7 +170,10 @@ if [ "$ENABLE_ALEMBIC_MIGRATIONS" = "true" ]; then
 fi
 
 echo "::group::Delete Test Infrastructure"
-kubectl delete --kustomize apps/${KUBERNETES_DEPLOYMENT_NAME}/test
+# skip deleting resources deployed by test-refinery-gatway
+if [ $KUBERNETES_DEPLOYMENT_NAME != "refinery-config" ] && [ $KUBERNETES_DEPLOYMENT_NAME != "refinery-websocket" ]; then
+    kubectl delete --kustomize apps/${KUBERNETES_DEPLOYMENT_NAME}/test
+fi
 kubectl delete --kustomize infrastructure/test
 echo "::endgroup::"
